fix: drop game_curses.c circular dependency
