+++ src/game_utils.h
@@ -297,3 +297,4 @@ foeTurnOption enemyTurnPick(Enemy * e, Fighter * f);
 foeTurnOption bossTurnPick(Boss * b, Fighter * f);
 void quit(Fighter * p, Room * room, loadInfo * load_info, Koliseo_Temp * t_kls);
 void setCounter(Turncounter * c, int turns);
+void useConsumable(Fighter * f, Enemy * e, Boss * b, char *string, int isBoss);
